T6876: DHCP-server increase retries that Kea makes to open a socket
KEA DHCP-server sometimes could be in the race condition when an
interface is not in the UP state.
The server tries to open the socket 5 times with an interval
of 5 seconds, then just starts as it is.
In this case, users cannot get leases from the interface, which was
in the DOWN state before starting KEA, but it is now in the UP state.

Increase 'service-sockets-max-retries' from 5 to 60