configd: T6747: use one long-lived instance of FRRender (#4274)
* smoketest: T6747: call wait after commit() only for FRR related tests

Commit 702a60a8de28 ("smoketest: T6746: wait after commit() until frr-reload
is no longer running") added a guard timeout for every commit executed via CLI
smoketests. This commit changes the bahavior to only add the guard timeout
for FRR related testscases.

This improves the overall smoketest time.

* configd: T6747: use one long-lived instance of FRRender

Previously there was one FRRender() instance per config session. This resulted
in re-rendering the FRR configuration every time a new config session was
created.

Example:

  vyos@vyos:~$ configure
  vyos@vyos# set interfaces dummy dum0 description foo
  vyos@vyos# commit
  vyos@vyos# exit

  vyos@vyos:~$ configure
  vyos@vyos# set interfaces dummy dum0 description bar
  vyos@vyos# commit
  vyos@vyos# exit

In the past this caused a re-render of the FRR configuration as the delta check
added in commit ec80c75d6776 ("frrender: T6746: only re-render FRR config if
config_dict did change") evaluated to false, as it operated on a new instance
of the FRRender class.

With this change there is no FRR re-render, as there is nothing to update
in FRR.