smoketest: T7023: add tac_plus container to live validate login (#4285)
* smoketest: T7023: unify container image loading

* smoketest: T7023: add tac_plus container to live validate login

TACACS is pretty sensible to its configuration. Instead of manual testing,
extend the smoketest platform to ship a tac_plus container and perform logins
against a locally running tac_plus server in a container.

The login username/password and TACACS shared secret is generated randomly on
the fly for every testcase.